refactor(echidna): extract shared environment controls
YClipControl, YLevelLock, DisplaySettings move to panels/controls/ so
they can be composed by both BuildPanel and the upcoming AnimateRightPanel
extension. BuildPanel behavior is unchanged.
